# backend/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "Nook API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:5999/"
        }
    ],
    "paths": {
...
        "/api/nodes/{id}": {
            "get": {
                "tags": [
                    "Nodes"
                ],
                "operationId": "GetNode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeDto"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Nodes"
                ],
                "operationId": "PatchNode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchNodeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Nodes"
                ],
                "operationId": "DeleteNode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/nodes/{id}/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "operationId": "GetNodeProperties",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonElement"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "description": "Replaces all page properties. Values are validated by type (dates must be ISO-8601).",
                "operationId": "SetNodeProperties",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonElement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonElement"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Properties"
                ],
                "description": "Merges the given properties; a null value removes that property.",
                "operationId": "PatchNodeProperties",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonElement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonElement"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/aliases": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "operationId": "GetNodeAliases",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "description": "Aliases are unique (case-insensitive) per workspace; 409 carries the conflicting node id.",
                "operationId": "SetNodeAliases",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetAliasesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JsonElement": {},
...
            "NodeCover": {
                "required": [
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "position": {
                        "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
                        "type": [
                            "null",
                            "number",
                            "string"
                        ],
                        "format": "double"
                    }
                }
            },
            "NodeDto": {
                "required": [
                    "id",
                    "workspaceId",
                    "parentId",
                    "kind",
                    "title",
                    "icon",
                    "cover",
                    "position",
                    "pageSettings",
                    "archivedAt",
                    "deletedAt",
                    "createdAt",
                    "updatedAt",
                    "effectiveRole"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "workspaceId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "parentId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    },
                    "kind": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "icon": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeIcon"
                            }
                        ]
                    },
                    "cover": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeCover"
                            }
                        ]
                    },
                    "position": {
                        "type": "string"
                    },
                    "pageSettings": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/PageSettings"
                            }
                        ]
                    },
                    "archivedAt": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "effectiveRole": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "hasChildren": {
                        "type": "boolean",
                        "default": false
                    },
                    "properties": {}
                }
            },
            "NodeIcon": {
                "required": [
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "OptionalOfGuid": {
                "type": [
                    "null",
                    "string"
                ],
                "format": "uuid"
            },
            "OptionalOfNodeCover": {
                "type": [
                    "null",
                    "object"
                ],
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "value": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "position": {
                        "type": [
                            "null",
                            "number"
                        ]
                    }
                }
            },
            "OptionalOfNodeIcon": {
                "type": [
                    "null",
                    "object"
                ],
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "value": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "PageSettings": {
                "type": "object",
                "properties": {
                    "font": {
                        "type": "string",
                        "default": "default"
                    },
                    "smallText": {
                        "type": "boolean",
                        "default": false
                    },
                    "fullWidth": {
                        "type": "boolean",
                        "default": false
                    },
                    "locked": {
                        "type": "boolean",
                        "default": false
                    }
                }
            },
            "PatchNodeRequest": {
                "required": [
                    "title",
                    "icon",
                    "cover",
                    "parentId",
                    "position",
                    "pageSettings"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "icon": {
                        "$ref": "#/components/schemas/OptionalOfNodeIcon"
                    },
                    "cover": {
                        "$ref": "#/components/schemas/OptionalOfNodeCover"
                    },
                    "parentId": {
                        "$ref": "#/components/schemas/OptionalOfGuid"
                    },
                    "position": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "pageSettings": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/PageSettings"
                            }
                        ]
                    }
                }
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "status": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "null",
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "detail": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "instance": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
            "SetAliasesRequest": {
                "required": [
                    "aliases"
                ],
                "type": "object",
                "properties": {
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    },
...
}